Restructure client
