+++ matter_server/client/__init__.py
@@ -1,4 +1,4 @@
 """Client for the MatterServer."""
-from .client import MatterClient
+from .matter import MatterClient
 
 __all__ = ["MatterClient"]
